Handle crashes that may happen before the main loop
