+++ main/include/crash_handlers.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+namespace hex::crash {
+
+    void setupCrashHandlers();
+}
